refactor: AdventureApiTest, AttendanceApiTest
