+++ src/main/java/com/playkuround/playkuroundserver/domain/common/BaseTimeEntity.java
@@ -1,13 +1,13 @@
 package com.playkuround.playkuroundserver.domain.common;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
 import java.time.LocalDateTime;
 
 @MappedSuperclass
@@ -24,3 +24,4 @@ public abstract class BaseTimeEntity {
     private LocalDateTime updatedAt;
 
 }
+
